+++ scripts/run-tests.mjs
@@ -1,5 +1,6 @@
 import '../src/test/app/simulation.js';
 import '../src/test/app/app.js';
+import '../src/test/bikes/autodetect.js';
 import '../src/test/bikes/bot.js';
 import '../src/test/bikes/flywheel.js';
 import '../src/test/bikes/ic4.js';